씬 구분
씬이 구분되도록 메인게임 클래스를 나누고 씬을 클래스별로 나눴습니다.
