+++ D3DX_PROJECT/cSCENE_RESULT.h
@@ -1,0 +1,16 @@
+#pragma once
+class cSCENE_RESULT
+{
+public:
+	cSCENE_RESULT();
+	~cSCENE_RESULT();
+
+	void Setup();
+	void Update();
+	void Render();
+	void WndProc(HWND hWnd, UINT message, WPARAM wParam, LPARAM lParam);
+
+private:
+	cCamera* m_pCamera;
+};
+
