+++ D3DX_PROJECT/cSCENE_TITLE.h
@@ -1,0 +1,16 @@
+#pragma once
+class cSCENE_TITLE
+{
+public:
+	cSCENE_TITLE();
+	~cSCENE_TITLE();
+
+	void Setup();
+	void Update();
+	void Render();
+	void WndProc(HWND hWnd, UINT message, WPARAM wParam, LPARAM lParam);
+
+private:
+	cCamera* m_pCamera;
+};
+
